% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{add_gen_liab_to_graph}
\alias{add_gen_liab_to_graph}
\title{Add pseudo point to graph}
\usage{
add_gen_liab_to_graph(fam_graph, index_id)
}
\arguments{
\item{fam_graph}{family graph center on proband}

\item{index_id}{id of proband.}
}
\value{
returns family graph with a pseudo point added that will have the same connections as the proband point. For kinship construction, the relationship to the proband must be adjusted to 1 \* h2 (and not 0.5 \* h2).
}
\description{
Internal function. Identifies the proband and makes a duplicated point that will act as genetic liability in kinship matrix
}
