% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{add_missing_roles_for_proband}
\alias{add_missing_roles_for_proband}
\title{Add in missing roles for proband}
\usage{
add_missing_roles_for_proband(
  temp_tbl,
  role,
  cur_roles,
  cur_fam_id,
  pid,
  fam_id,
  phen_names = NULL
)
}
\arguments{
\item{temp_tbl}{tibble to add missing proband roles to; originates from .tbl of estimate_liability.}

\item{role}{name of role column}

\item{cur_roles}{values of the role column}

\item{cur_fam_id}{current family ID being worked on}

\item{pid}{name of column with personal IDs}

\item{fam_id}{name of column with family IDs}

\item{phen_names}{vector of phenotype names as given in .tbl of estimate_liability. Defaults to NULL (which is single trait).}
}
\value{
The provided temp_tbl object is returned, but with the missing "g" and/or "o" roles added, where -Inf and Inf values
have been used to fill the lower and upper threshold values. If phen_names is provided, a pair of upper and lower values is
provided for each entry in phen_names.
}
\description{
This function adds missing roles for a proband and fills lower threshold values with -Inf and upper threshold values with Inf.
}
