% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cir_functions.R
\name{age_to_cir}
\alias{age_to_cir}
\title{Calculate cumulative incidence rate based off of a person's age (Used for Examples)}
\usage{
age_to_cir(age, pop_prev = 0.1, age_mid = 60, slope = 1/8)
}
\arguments{
\item{age}{age of individual.}

\item{pop_prev}{Overall prevalence in the population.}

\item{age_mid}{The mid point of the logistic function}

\item{slope}{The rate of increase.}
}
\value{
Returns the estimated genetic liabilities.
}
\description{
Calculate cumulative incidence rate based off of a person's age (Used for Examples)
}
\examples{
# See R/Example/example_nosib.R for an example of use and input.
# curve(age_to_cir, from = 10, to = 110, xname = "age")
}
