% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cir_functions.R
\name{age_to_thres_trunc_normal}
\alias{age_to_thres_trunc_normal}
\title{calculate the thresholds of an individual based off of an individual's age using the truncated normal distribution}
\usage{
age_to_thres_trunc_normal(
  age,
  min_age = 10,
  max_age = 100 - min_age,
  a = stats::qnorm(0.05, lower.tail = F),
  b = Inf
)
}
\arguments{
\item{age}{true underlying liability for an individual}

\item{min_age}{minimum age to consider}

\item{max_age}{maximum age to consider}

\item{a}{the lower cutoff point of the truncated normal distribution}

\item{b}{the upper cutoff point of the truncated normal distribution}
}
\value{
Returns the estimated genetic liabilities.
}
\description{
calculate the thresholds of an individual based off of an individual's age using the truncated normal distribution
}
\examples{
# See R/Example/example_nosib.R for an example of use and input.
# curve(age_to_thres_trunc_normal, from = 10, to = 100, xname = "age")
}
