% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{check_correlation_matrix}
\alias{check_correlation_matrix}
\title{Checking that a correlation matrix is valid}
\usage{
check_correlation_matrix(corrmat)
}
\arguments{
\item{corrmat}{A numeric matrix holding the correlations.
All diagonal entries must be equal to one, while all off-diagonal entries
must be between -1 and 1. In addition, the matrix must be symmetric.}
}
\value{
If \code{corrmat} is a valid correlation matrix that is symmetric,
has one on all diagonal entries and numbers between -1 and 1 on all off-
diagonal entries,
then the function will return TRUE. Otherwise, the function will be aborted.
}
\description{
\code{check_correlation_matrix} checks whether a matrix is a valid
correlation matrix, i.e. whether its diagonal entries are equal to one,
while all off-diagonal entries are between -1 and 1, and whether it is
symmetric.
}
\details{
This function can be used to check whether a correlation matrix has diagonal
entries equal to 1 and off-diagonal entries between -1 and 1 as well as whether
it is symmetric.
}
\examples{
check_correlation_matrix(matrix(c(1,0.4,0.4,1), nrow = 2))
check_correlation_matrix(diag(3))


\dontrun{
# This will result in errors:
check_correlation_matrix(matrix(c(0.2,0.4,0.4,0.2), nrow = 2))
check_correlation_matrix(matrix(nrow=2, ncol = 2))
}

}
