% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments_functions.R
\name{check_family_role}
\alias{check_family_role}
\title{Check that the strings are valid for family members}
\usage{
check_family_role(s1)
}
\arguments{
\item{s1}{String representing the family member.
The string must be one of the following:
\itemize{
\item g (Genetic component of full liability)
\item o (Full liability)
\item m (Mother)
\item f (Father)
\item mgm (Maternal grandmother)
\item mgf (Maternal grandfather)
\item mgp (Maternal grandparent)
\item pgm (Paternal grandmother)
\item pgf (Paternal grandfather)
\item pgp (Paternal grandparent)
\item s[0-9]* (Full siblings)
\item c[0-9]* (children)
\item mhs[0-9]* (Half-siblings - maternal side)
\item phs[0-9]* (Half-siblings - paternal side)
\item mau[0-9]* (Aunts/Uncles - maternal side)
\item pau[0-9]* (Aunts/Uncles - paternal side).
}}
}
\value{
If s1 is a valid string from the mentioned list of strings no errors will be returned.
}
\description{
\code{check_family_role} checks that the strings used for the relatedness times the
heritability for a pair of family members is valid
}
\examples{
check_family_role("g")
check_family_role("f")
check_family_role("s")

\dontrun{
# This will result in errors:
check_family_role("a")
check_family_role(mhs)
}

}
