% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_functions.R
\name{check_positive_definite}
\alias{check_positive_definite}
\title{Verifies that the covariance matrix is positive definite (a properti of gaussian covariance matrices) by checking if all eigen values are positive.}
\usage{
check_positive_definite(full_cov, corr_mat, correction_val = 0.99, n_sib = 0)
}
\arguments{
\item{full_cov}{The full covariance matrix for the LT-FH++ method.}

\item{corr_mat}{Matrix containing the heritabilities and correlations for each of the phenotypes provided. Heritabilities should be on the diagonal with the corresponing off-diagonal entry holding the correlation between two phenotypes.}

\item{correction_val}{value to multiply to off-diagonal entries in an attempt to force positive definite.}

\item{n_sib}{Number of Siblings to include in the matrix.}
}
\value{
Returns the full covariance matrix based off of the heritabilities and corralations provided in corr_mat.
}
\description{
Verifies that the covariance matrix is positive definite (a properti of gaussian covariance matrices) by checking if all eigen values are positive.
}
\examples{
check_positive_definite(get_cov(.5), matrix(.5, 1, 1))

}
