% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{check_proportion}
\alias{check_proportion}
\title{Checking that proportions are valid}
\usage{
check_proportion(prop)
}
\arguments{
\item{prop}{A number, integer or numeric vector representing the proportions that
need to be validated.}
}
\value{
If \code{prop} are valid proportions of class \code{numeric}
or \code{integer}, non-negative and at most one,
then the function will return TRUE. Otherwise, the function aborts.
}
\description{
\code{check_proportion} checks whether proportions are valid, i.e.
whether they are non-negative and at most one.
}
\details{
This function can be used to check whether proportions are non-negative
and at most one.
}
\examples{
\dontrun{
check_proportion(0.2)
check_proportion(0.04)
check_proportion(0)
check_proportion(1)

# This will result in errors:
check_proportion(2)
check_proportion(-0.5)
}

}
