% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{check_valid_relatives}
\alias{check_valid_relatives}
\title{Checking that relatives are represented by valid strings}
\usage{
check_valid_relatives(relatives)
}
\arguments{
\item{relatives}{A string or character vector representing
the relatives.
All strings must be chosen among the following abbreviations
\itemize{
\item g (Genetic component of full liability)
\item o (Full liability)
\item m (Mother)
\item f (Father)
\item mgm (Maternal grandmother)
\item mgf (Maternal grandfather)
\item pgm (Paternal grandmother)
\item pgf (Paternal grandfather)
\item s[0-9]* (Full siblings)
\item mhs[0-9]* (Half-siblings - maternal side)
\item phs[0-9]* (Half-siblings - paternal side)
\item mau[0-9]* (Aunts/Uncles - maternal side)
\item pau[0-9]* (Aunts/Uncles - paternal side)
for the function to return TRUE.
}}
}
\value{
If \code{relatives} is a string or character vector such that
all strings are chosen from the mentioned list of strings,
then the function will return TRUE. Otherwise, the function is aborted.
}
\description{
\code{check_valid_relatives} checks whether relatives are represented
by valid abbreviations.
}
\details{
This function can be used to check whether relatives are represented
by valid abbreviations. A valid abbreviation is one of the following:
\itemize{
\item g (Genetic component of full liability)
\item o (Full liability)
\item m (Mother)
\item f (Father)
\item mgm (Maternal grandmother)
\item mgf (Maternal grandfather)
\item pgm (Paternal grandmother)
\item pgf (Paternal grandfather)
\item s[0-9]* (Full siblings)
\item mhs[0-9]* (Half-siblings - maternal side)
\item phs[0-9]* (Half-siblings - paternal side)
\item mau[0-9]* (Aunts/Uncles - maternal side)
\item pau[0-9]* (Aunts/Uncles - paternal side).
}
}
\examples{
\dontrun{
check_valid_relatives("g")
check_valid_relatives("o")
check_valid_relatives("mgm")

# This will result in errors:
get_relatedness("a")
get_relatedness(m)
}

}
