% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_under_LTM.R
\name{construct_aoo}
\alias{construct_aoo}
\title{Constructing age of onset (aoo)}
\usage{
construct_aoo(fam_mem, .tbl, pop_prev)
}
\arguments{
\item{fam_mem}{A character vector holding all family members.}

\item{.tbl}{A tibble holding the liability as well as age and
disease status for the set of individuals in fam_mem.}

\item{pop_prev}{A positive number representing the population prevalence, i.e. the
overall prevalence in the population.}
}
\value{
A tibble holding all columns present in .tbl as well
as the age of onset or the current age
(depending on the disease status) for all individuals
given in fam_mem.
}
\description{
\code{construct_aoo} constructs the age of onset (aoo)
for a variable number of family members based on their
liability, disease status and current age.
}
