% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_under_LTM.R
\name{construct_thresholds}
\alias{construct_thresholds}
\title{Computing thresholds}
\usage{
construct_thresholds(fam_mem, .tbl, pop_prev)
}
\arguments{
\item{fam_mem}{A character vector holding all family members.}

\item{.tbl}{A tibble holding the family ID, disease status as well
as the age of onset or the current age
(depending on the disease status).}

\item{pop_prev}{A positive number representing the population prevalence, i.e. the
overall prevalence in the population.}
}
\value{
A tibble holding the personal identifier (PID) as well as
the lower and the upper threshold for all individuals
present in fem_mem.
}
\description{
\code{construct_thresholds} computes the upper and lower
thresholds for a variable number of family members based on their
disease status and current age or age of onset (depending on
the disease status).
}
