% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{correct_positive_definite_simplified}
\alias{correct_positive_definite_simplified}
\title{A simplied version of correct_positive_definite}
\usage{
correct_positive_definite_simplified(
  covmat,
  correction_limit = 100,
  correction_val = 0.99
)
}
\arguments{
\item{covmat}{Covariance matrix that needs to be positive definite}

\item{correction_limit}{The maximum number of times to correct off-diagonal elements}

\item{correction_val}{The value to multiply on the off-diagnoal elements.}
}
\value{
Returns list with the corrected covmat and number of iterations used to achieve positive definite.
}
\description{
multiplies off-diagonal elements of a matrix with a correction value until all eigen values are positive or correction limit is reached.
}
