% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cir_functions.R
\name{est_cir}
\alias{est_cir}
\title{Estimating the cumulative incidence rate in sample}
\usage{
est_cir(
  data,
  indivs = c("child", "father", "mother"),
  ids = c("FID", "pid_f", "pid_m")
)
}
\arguments{
\item{data}{tibble containing the status column, age, and id of individuals in "indivs" and "ids.}

\item{indivs}{vector of prefixes to construct status and age columns from}

\item{ids}{unique ids for every individual in indivs. These ids will be the identifier for a given individual in the output.}
}
\value{
Returns cir for the provided data. Estimated in-sample.
}
\description{
Estimating the cumulative incidence rate in sample
}
