% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_gen_liab_multiple_traits.R
\name{estimate_gen_liability_multi_trait}
\alias{estimate_gen_liability_multi_trait}
\title{Estimates genetic liability based off of multiple phenotypes, family history, and temportal information.}
\usage{
estimate_gen_liability_multi_trait(
  corr_mat,
  phen.list,
  thr.list,
  ind = c(1),
  id_col = "ids",
  tol = 0.01
)
}
\arguments{
\item{corr_mat}{Correlation matrix for each of the traits being analysed.}

\item{phen.list}{list of tibble or data.frame with the IDs and status of the genotyped individual and the parents and siblings.}

\item{thr.list}{list of tibble or data.frame with a row of each individual in the provided phen. Each row should contain the ID and threshold value needed for the model.}

\item{ind}{Indices to return from the gibbs sampler. 1 corresponds to the genetic liability of the first phenotype. c(1,5) corresponds to the genetic liability of the first two phenotypes provided, but with no siblings in the model.}

\item{id_col}{Column names of IDs for family members.}

\item{tol}{Convergence criteria of the gibbs sampler. Default is 0.01, meaning a standard error of the mean below 0.01}
}
\value{
Returns the estimated genetic liabilities.
}
\description{
Estimates genetic liability based off of multiple phenotypes, family history, and temportal information.
}
\examples{
# See R/Example/example_multi.R for an example of use and input.

}
