% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_gen_liab_noFH_corr_trait.R
\name{estimate_gen_liability_noFH_corr_traits}
\alias{estimate_gen_liability_noFH_corr_traits}
\title{Estimate genetic liability when multiple traits and prevalence information for each trait are available, but family history is not}
\usage{
estimate_gen_liability_noFH_corr_traits(
  phen,
  h2_vec,
  gen_cor_vec,
  prev_vec,
  tol = 0.01
)
}
\arguments{
\item{phen}{tibble or data.frame with status of the individual in order of h2_vec.}

\item{h2_vec}{Liability scale heritability of the trait being analysed.}

\item{gen_cor_vec}{vector of genetic correlations, in order of appearance in phenotype file}

\item{prev_vec}{Vector of prevalences in order of appearance in h2_vec.}

\item{tol}{Convergence criteria of the Gibbs sampler. Default is 0.01, meaning a standard error of the mean below 0.01}
}
\value{
Returns the estimated genetic liabilities.
}
\description{
Estimate genetic liability when multiple traits and prevalence information for each trait are available, but family history is not
}
\examples{
# See R/Example/example_nosib.R for an example of use and input.
}
