% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_gen_liab_noFH_corr_trait.R
\name{estimate_liability_prevalence}
\alias{estimate_liability_prevalence}
\title{Estimating the genetic or full liability for multiple phenotypes
using prevalence information}
\usage{
estimate_liability_prevalence(
  status,
  h2_vec,
  genetic_corrmat,
  full_corrmat,
  prevalences,
  pid = "PID",
  out = c(1),
  tol = 0.01,
  parallel = FALSE,
  progress = FALSE
)
}
\arguments{
\item{status}{A matrix, list or data frame that can be converted into a tibble.
Must have at least three columns; one holding the personal identifier for all individuals,
and the remaining holding the phenotype status for the first and second
phenotype, respectively. It must be possible to tie each status variable
to a specific phenotype uniquely. The function will use the column names to create
phenotype names.}

\item{h2_vec}{A numeric vector representing the heritabilites on liability scale
for all phenotypes. All entries in \code{h2_vec} must be non-negative and at most 1.}

\item{genetic_corrmat}{A numeric matrix holding the genetic correlations between the desired
phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries
must be between -1 and 1. In addition, the matrix must be symmetric.
Defaults to NULL.}

\item{full_corrmat}{A  numeric matrix holding the full correlations between the desired
phenotypes. All diagonal entries must be equal to one, while all off-diagonal entries
must be between -1 and 1. In addition, the matrix must be symmetric.
Defaults to NULL.}

\item{prevalences}{A numeric, non-negative vector holding the prevalences.
All prevalences must be at most one.}

\item{pid}{A string holding the name of the column in \code{status} that hold
the personal identifier. Defaults to "PID".}

\item{out}{A character or numeric vector indicating whether the genetic component
of the full liability, the full liability or both should be returned. If out = c(1) or
out = c("genetic"), the genetic liability is estimated and returned. If out = c(2) or
out = c("full"), the full liability is estimated and returned. If out = c(1,2) or
out = c("genetic", "full"), both components are estimated and returned.
Defaults to c(1).}

\item{tol}{A number that is used as the convergence criterion for the Gibbs sampler.
Equals the standard error of the mean. That is, a tolerance of 0.2 means that the
standard error of the mean is below 0.2. Defaults to 0.01.}

\item{parallel}{A logical scalar indicating whether computations should be performed parallel.
In order for this to be possible, the user must install the library "future.apply" and create a plan
(see \code{\link[future.apply]{future_apply}}). Defaults to FALSE.}

\item{progress}{A logical scalar indicating whether the function should display
a progress bar. Defaults to \code{FALSE}.}
}
\value{
If \code{status} is a matrix, list or data frame that can be converted into a
tibble and that has a column named \code{PID} and if the heritabilities, corrmat,
out and tol are of the required form, then the function returns a tibble with at least
five columns (depending on the length of out).
The first column corresponds to the columns pid.
If out is equal to c(1) or c("genetic"), the second and third columns hold the estimated genetic
liability as well as the corresponding standard error for the first phenotype, respectively.
If out equals c(2) or c("full"), the second and third columns hold the estimated full liability
as well as the corresponding standard error for the first phenotype, respectively.
If out is equal to c(1,2) or c("genetic","full"), the second and third columns hold the estimated
genetic liability as well as the corresponding standard error for the first phenotype, while the fourth and
fifth columns hold the estimated full liability as well as the corresponding standard error for the
same phenotype.
The remaining columns hold the estimated genetic liabilities and/or the estimated full liabilities
as well as the corresponding standard errors for the remaining phenotypes.
}
\description{
FUNCTIONS FOR MULTIPLE TRAITS IS STILL NOT COMPLETE. PLEASE DO NOT USE.
}
\details{
\code{estimate_liability_prevalence} estimates the genetic component of
the full liability and/or the full liability for a number of individuals
based solely on prevalence information for a variable number of
phenotypes.

This function can be used to estimate either the genetic component of the
full liability, the full liability or both for a variable number of traits,
when no family history is available, but prevalence information for each
phenotype can be obtained.
}
\seealso{
\code{\link[future.apply]{future_apply}}
}
