% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_liabilities.R
\name{estimate_liability_single}
\alias{estimate_liability_single}
\title{Estimating the genetic or full liability}
\usage{
estimate_liability_single(
  family,
  threshs,
  h2 = 0.5,
  pid = "PID",
  fam_id = "fam_ID",
  out = c(1),
  tol = 0.01,
  always_add = c("g", "o"),
  progress = FALSE
)
}
\arguments{
\item{family}{A matrix, list or data frame that can be converted into a tibble.
Must have at least two columns that hold the family identifier and the corresponding
personal identifiers, respectively. That is, for each family in \code{fam_id} there should
be a list holding all individuals belonging to that family in \code{pid}. Note that the
personal identifiers for all individuals must have a special format. It must end
on \code{_?}, where \code{?} is one of the following abbreviations
\itemize{
\item g (Genetic component of the full liability)
\item o (Full liability)
\item m (Mother)
\item f (Father)
\item c[0-9]\\emph{.[0-9]\}(Children)
\item mgm (Maternal grandmother)
\item mgf (Maternal grandfather)
\item pgm (Paternal grandmother)
\item pgf (Paternal grandfather)
\item s[0-9]* (Full siblings)
\item mhs[0-9]* (Half-siblings - maternal side)
\item phs[0-9]* (Half-siblings - paternal side)
\item mau[0-9]* (Aunts/Uncles - maternal side)
\item pau[0-9]* (Aunts/Uncles - paternal side).
See also \code{\link{construct_covmat}}.
}}

\item{threshs}{A matrix, list or data frame that can be converted into a tibble.
Must have at least three columns, one holding the personal identifier for all individuals,
and the remaining two holding the lower and upper thresholds, respectively. The latter
should be called "lower" and "upper", respectively.}

\item{h2}{A number representing the heritability on liability scale
for a single phenotype. Must be non-negative. Note that under the liability threshold model,
the heritability must also be at most 1.
Defaults to 0.5.}

\item{pid}{A string holding the name of the column in \code{family} and
\code{threshs} that hold the personal identifier(s). Defaults to "PID".}

\item{fam_id}{A string holding the name of the column in \code{family} that
holds the family identifier. Defaults to "fam_ID".}

\item{out}{A character or numeric vector indicating whether the genetic component
of the full liability, the full liability or both should be returned. If \code{out = c(1)} or
\code{out = c("genetic")}, the genetic liability is estimated and returned. If \code{out = c(2)} or
\code{out = c("full")}, the full liability is estimated and returned. If \code{out = c(1,2)} or
\code{out = c("genetic", "full")}, both components are estimated and returned.
Defaults to \code{c(1)}.}

\item{tol}{A number that is used as the convergence criterion for the Gibbs sampler.
Equals the standard error of the mean. That is, a tolerance of 0.2 means that the
standard error of the mean is below 0.2. Defaults to 0.01.}

\item{always_add}{A character vector or \code{NULL}. If \code{always_add = c("g","o")}, both the genetic component
of the full liability as well as the full liability will be added to the list of family members.
If always_add equals \code{"g"} or \code{"o"}, the genetic component of the full liability or the full liability
will be added, respectively. If \code{always_add = NULL}, no component will be added.
Defaults to \code{c("g","o")}.}

\item{progress}{A logical scalar indicating whether the function should display
a progress bar. Defaults to \code{FALSE}.}
}
\value{
If \code{family} and \code{threshs} are two matrices, lists or
data frames that can be converted into tibbles, if \code{family} has two
columns named like the strings represented in \code{pid} and \code{fam_id}, if
\code{threshs} has a column named like the string given in \code{pid} as
well as a column named "lower" and a column named "upper" and if the
liability-scale heritability \code{h2}, \code{out}, \code{tol} and
\code{always_add} are of the required form, then the function returns a
tibble with either four or six columns (depending on the length of out).
The first two columns correspond to the columns \code{fam_id} and \code{pid} '
present in \code{family}.
If \code{out} is equal to \code{c(1)} or \code{c("genetic")}, the third
and fourth column hold the estimated genetic liability as well as the
corresponding standard error, respectively.
If \code{out} equals \code{c(2)} or \code{c("full")}, the third and
fourth column hold the estimated full liability as well as the
corresponding standard error, respectively.
If \code{out} is equal to \code{c(1,2)} or \code{c("genetic","full")},
the third and fourth column hold the estimated genetic liability as
well as the corresponding standard error, respectively, while the fifth and
sixth column hold the estimated full liability as well as the corresponding
standard error, respectively.
}
\description{
\code{estimate_liability_single} estimates the genetic component of the full
liability and/or the full liability for a number of individuals based
on their family history.
}
\details{
This function can be used to estimate either the genetic component of the
full liability, the full liability or both.
}
\examples{
sims <- simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, 
add_ind = TRUE, h2 = 0.5, n_sim=10, pop_prev = .05)

estimate_liability_single(family = sims$fam_ID, threshs = sims$thresholds, 
h2 = 0.5, pid = "indiv_ID", fam_id = "fam_ID", out = c(1), tol = 0.01,
always_add = c("g","o"), progress = TRUE)
# 
sims <- simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = TRUE, 
h2 = 0.5, n_sim=10, pop_prev = .05)

estimate_liability_single(family = sims$fam_ID, threshs = sims$thresholds, 
h2 = 0.5, pid = "indiv_ID", fam_id = "fam_ID", out = c("genetic"), 
tol = 0.01, always_add = c("g","o"))

}
\seealso{
\code{\link[future.apply]{future_apply}}
}
