% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_to_trio.R
\name{fixSexCoding}
\alias{fixSexCoding}
\title{Fixing sex coding in trio info}
\usage{
fixSexCoding(x, sex_coding = TRUE, dadid, momid)
}
\arguments{
\item{x}{current row to check against}

\item{sex_coding}{logical. Is sex coded as character?}

\item{dadid}{column name of father ids}

\item{momid}{column name of mother ids}
}
\value{
appropriate sex coding
}
\description{
Internal function used to assist in fixing sex coding separately from id coding type.
}
