% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_functions.R
\name{generate_corr_matrix}
\alias{generate_corr_matrix}
\title{Constructs the correlation matrix}
\usage{
generate_corr_matrix(
  names,
  heri,
  corr,
  heritability_col = "h2 (Liability Scale)"
)
}
\arguments{
\item{names}{Names of disorders to include}

\item{heri}{list, data.frame, or tibble with the heritabilities stored under the provded names}

\item{corr}{similar, but for correlations.}

\item{heritability_col}{Name of column that contains the heritabilities.}
}
\value{
Returns the correlation matrix needed as input for get_full_cov(), the LT-FH++ method and similar.
}
\description{
Constructs the correlation matrix
}
\examples{

}
