% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{get_all_combs}
\alias{get_all_combs}
\title{construct all combinations of input vector}
\usage{
get_all_combs(vec)
}
\arguments{
\item{vec}{vector of strings}
}
\value{
A vector of strings is returned.
}
\description{
pastes together all combinations of input vector
}
\examples{
\dontrun{
get_all_combs(letters[1:3])
}


}
