% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_covariance_functions.R
\name{get_between_trait_cov}
\alias{get_between_trait_cov}
\title{Helps construct the correlation entries in the covariance matrix needed for LT-FH++ with multiple traits.}
\usage{
get_between_trait_cov(h2_1, h2_2, rho, n_sib = 0)
}
\arguments{
\item{h2_1}{Heritability estimate on liability scale for the first phenotype to construct the correlation entries of the covariance matrix off of.}

\item{h2_2}{Heritability estimate on liability scale for the second phenotype to construct the correlation entries of the covariance matrix off of.}

\item{rho}{Correlation between the two phenotypes.}

\item{n_sib}{Number of Siblings to include in the matrix.}
}
\description{
This function has been retired in favor of \code{\link{construct_covmat}}. This function is kept for legacy reasons.
}
\examples{
get_between_trait_cov(.5, .5, .5)

}
