% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_functions.R
\name{get_cov}
\alias{get_cov}
\title{Creates the covariance matrix needed for LT-FH++}
\usage{
get_cov(h2, n_sib = 0)
}
\arguments{
\item{h2}{Heritability estimate on liability scale to construct the covariance matrix off of.}

\item{n_sib}{Number of Siblings to include in the covariance matrix.}
}
\value{
None
}
\description{
Creates the covariance matrix needed for LT-FH++
}
\examples{
get_cov(.5)

}
