% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{get_kinship}
\alias{get_kinship}
\title{Construct kinship matrix from graph}
\usage{
get_kinship(fam_graph, h2, index_id = NA, add_ind = TRUE)
}
\arguments{
\item{fam_graph}{graph.}

\item{h2}{heritability.}

\item{index_id}{proband id. Only used in conjuction with add_ind = TRUE.}

\item{add_ind}{add genetic liability to the kinship matrix. Defaults to true.}
}
\value{
A kinship matrix.
}
\description{
construct the kinship matrix from a graph representation of a family, centered on an index person (proband).
}
