% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cir_functions.R
\name{liab_to_aoo}
\alias{liab_to_aoo}
\title{Calculate a person's age of onset based off of the true underlying liability (Used for Examples)}
\usage{
liab_to_aoo(liab, pop_prev = 0.1, age_mid = 60, slope = 1/8)
}
\arguments{
\item{liab}{true underlying liability of an individual.}

\item{pop_prev}{Overall prevalence in the population.}

\item{age_mid}{The mid point of the logistic function}

\item{slope}{The rate of increase.}
}
\value{
Returns the estimated genetic liabilities.
}
\description{
Calculate a person's age of onset based off of the true underlying liability (Used for Examples)
}
\examples{
# See R/Example/example_nosib.R for an example of use and input.
# curve(liab_to_aoo, from = 1.3, to = 3.5, xname = "liab")
}
