% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cir_functions.R
\name{liab_to_aoo_case_trunc_normal}
\alias{liab_to_aoo_case_trunc_normal}
\title{calculate the age of onset for cases based off of the true underlying liability using the truncated normal distribution}
\usage{
liab_to_aoo_case_trunc_normal(
  liab,
  min_aoo = 10,
  max_aoo = 100 - min_aoo,
  a = stats::qnorm(0.05, lower.tail = F),
  b = Inf
)
}
\arguments{
\item{liab}{true underlying liability for an individual}

\item{min_aoo}{earliest age of onset to consider}

\item{max_aoo}{latest age of onset to consider}

\item{a}{the lower cutoff point of the truncated normal distribution}

\item{b}{the upper cutoff point of the truncated normal distribution}
}
\value{
Returns the estimated genetic liabilities.
}
\description{
calculate the age of onset for cases based off of the true underlying liability using the truncated normal distribution
}
\examples{
# See R/Example/example_nosib.R for an example of use and input.
# curve(liab_to_aoo_case_trunc_normal, from = qnorm(0.05, lower.tail = F), to = 3.5, xname = "liab")
}
