% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFHPlus-package.R
\docType{package}
\name{LTFHPlus-package}
\alias{LTFHPlus}
\alias{LTFHPlus-package}
\title{LTFHPlus: Implementation of LT-FH++}
\description{
Implementation of LT-FH++, an extension of the liability threshold family history (LT-FH) model. LT-FH++ uses a Gibbs sampler for sampling from the truncated multivariate normal distribution and allows for flexible family structures. LT-FH++ was first described in Pedersen, Emil M., et al. (2022) \doi{10.1016/j.ajhg.2022.01.009} as an extension to LT-FH with more flexible family structures, and again as the age-dependent liability threshold (ADuLT) model Pedersen, Emil M., et al. (2023) \url{https://www.nature.com/articles/s41467-023-41210-z} as an alternative to traditional time-to-event genome-wide association studies, where family history was not considered.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emilmip.github.io/LTFHPlus/}
  \item \url{https://github.com/EmilMiP/LTFHPlus}
  \item Report bugs at \url{https://github.com/EmilMiP/LTFHPlus/issues}
}

}
\author{
\strong{Maintainer}: Emil Michael Pedersen \email{emp@ph.au.dk}

Authors:
\itemize{
  \item Florian Privé \email{florian.prive.21@gmail.com} [thesis advisor]
  \item Jette Steinbach \email{jst@econ.au.dk}
}

Other contributors:
\itemize{
  \item Bjarni Jóhann Vilhjálmsson \email{bjv@econ.au.dk} [thesis advisor]
  \item Esben Agerbo \email{ea@econ.au.dk} [thesis advisor]
  \item Lucas Rasmussen \email{lar.ncrr@au.dk} [contributor]
}

}
\keyword{internal}
