% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFHpp-package.R
\docType{package}
\name{LTFHpp-package}
\alias{LTFHpp}
\alias{LTFHpp-package}
\title{LTFHpp: Extending the package 'LTFHPlus'}
\description{
Extending the available package 'LTFHPlus' in order to allow for more flexible family strctures. 'LT-FH++' is an extension of 'LT-FH' using a Gibbs sampler for sampling from the truncated multivariate normal distribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JetteS/LTFHPlus}
}

}
\author{
\strong{Maintainer}: Jette Steinbach \email{jst@econ.au.dk}

Authors:
\itemize{
  \item Emil Michael Pedersen \email{emp@ph.au.dk}
  \item Florian Privé \email{florian.prive.21@gmail.com}
}

}
\keyword{internal}
