% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_LTFHPlus_input.R
\name{prepare_graph}
\alias{prepare_graph}
\title{Construct graph from register information}
\usage{
prepare_graph(.tbl, fcol, mcol, icol, missingID_patterns = "0")
}
\arguments{
\item{.tbl}{tibble with columns icol, fcol, mcol. Additional columns will be attributes in the constructed graph.}

\item{fcol}{column name of column with father ids.}

\item{mcol}{column name of column with mother ids.}

\item{icol}{column name of column with proband ids.}

\item{missingID_patterns}{string of missing values in the ID columns. Multiple values can be used, but must be separated by "|". Defaults to "0".}
}
\value{
An igraph object. A (directed) graph object based on the links provided in .tbl.
}
\description{
\code{prepare_graph} constructs a graph based on mother, father, and offspring links.
}
