% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-sampler.R
\name{rtmvnorm.gibbs}
\alias{rtmvnorm.gibbs}
\title{Gibbs Sampler for the truncated multivariate normal distribution}
\usage{
rtmvnorm.gibbs(n_sim, sigma, lower, upper, fixed, ind = 1, burn_in = 1000)
}
\arguments{
\item{n_sim}{number of draws after burn in.}

\item{sigma}{covariance matrix for the normal distribution}

\item{lower}{lower truncation for the normal distribution. It can be -Inf}

\item{upper}{upper truncation for the normal distribution. It can be Inf}

\item{fixed}{Are any entries fixed ? Logical vector with an entry for each dimension}

\item{ind}{Indices to return from the gibbs sampler. 1 corresponds to the genetic liability of the first phenotype. c(1,5) corresponds to the genetic liability of the first two phenotypes provided, but with no siblings in the model.}

\item{burn_in}{Number of iterations that count as burn in for the gibbs sampler.}
}
\value{
Returns the estimated genetic liabilities.
}
\description{
Gibbs Sampler for the truncated multivariate normal distribution
}
