% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtmvnorm_gibbs_cpp}
\alias{rtmvnorm_gibbs_cpp}
\title{Gibbs Sampler for the truncated Normal distribution.}
\usage{
rtmvnorm_gibbs_cpp(P, sd, lower, upper, fixed, to_return, x, n_sim, burn_in)
}
\arguments{
\item{P}{conditional covariance matrix}

\item{sd}{covariance matrix}

\item{lower}{lower limit}

\item{upper}{upper limit}

\item{fixed}{logical vector, which entries are fixed?}

\item{to_return}{Which entries to return}

\item{x}{#starting value}

\item{n_sim}{number of simulations to perform after the burn in period}

\item{burn_in}{burn in period}
}
\description{
Gibbs Sampler for the truncated Normal distribution.
}
