% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_under_LTM.R
\name{simulate_under_LTM}
\alias{simulate_under_LTM}
\title{Simulate under the liability threshold model.}
\usage{
simulate_under_LTM(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  sq.herit = 0.5,
  n_sim = 1000,
  pop_prev = 0.1
)
}
\arguments{
\item{fam_vec}{A vector of strings holding the different
family members. All family members must be represented by strings from the
following list:
\itemize{
\item m (Mother)
\item f (Father)
\item mgm (Maternal grandmother)
\item mgf (Maternal grandfather)
\item pgm (Paternal grandmother)
\item pgf (Paternal grandfather)
\item s[0-9]* (Full siblings)
\item mhs[0-9]* (Half-siblings - maternal side)
\item phs[0-9]* (Half-siblings - paternal side)
\item mau[0-9]* (Aunts/Uncles - maternal side)
\item pau[0-9]* (Aunts/Uncles - paternal side).
Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").
}}

\item{n_fam}{A named vector holding the desired number of family members.
All names must be picked from the list mentioned above. Defaults to NULL.}

\item{add_ind}{A logical scalar indicating whether the genetic
component of the full liability as well as the full
liability for the underlying individual should be included in
the covariance matrix. Defaults to TRUE.}

\item{sq.herit}{A number representing the squared heritability on liability scale
for a single phenotype. Must be non-negative and at most 1.
Defaults to 0.5.}

\item{n_sim}{A positive number representing the number of simulations. Defaults to 1000.}

\item{pop_prev}{A positive number representing the population prevalence, i.e. the
overall prevalence in the population. Must be smaller than 1. Defaults to 0.1.}
}
\value{
If either fam_vec or n_fam is used as the argument, if it is of the required format,
if sq.herit is a number satisfying 0 <= sq.herit <= 1, n_sim is a strictly positive number,
and pop_prev is a positive number that is at most one,
then the output will be a list containing three tibbles. The first tibble, \code{sim_obs},
holds the disease status and the current age/age of onset for all family members in each
of the \code{n_sim} families. The second tibble, \code{thresholds}, holds the lower and upper thresholds
for all individuals in all families. Note that this tibble has the format required in
\code{\link{estimate_liability}}. The last tibble, \code{fam_ID}, connects the personal identifiers
for all individuals to the family identifiers. As \code{thresholds}, \code{fam_ID} has the format
required in \code{\link{estimate_liability}}.
Note that if neither fam_vec nor n_fam are specified, the function returns the disease status,
the current age/age of onset, the lower and upper thresholds, as well as the personal identifier
for a single individual, namely the individual under consideration (called 'o').
If both fam_vec and n_fam are defined, the user is asked to decide on which
of the two vectors to use.
}
\description{
\code{simulate_under_LTM} simulates thresholds under
the liability threshold model for a given family
structure.
}
\examples{
simulate_under_LTM()
simulate_under_LTM(fam_vec = NULL, n_fam = setNames(c(1,1,1,2,2), c("m","mgm","mgf","s","mhs")))
simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = F, sq.herit = 0.5, n_sim=500, pop_prev = .05)
simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = T, sq.herit = 0.5, n_sim = 200, pop_prev = 0.05)

}
\seealso{
\code{\link{construct_covmat}}
}
