% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cir_functions.R
\name{trunc_normal_cdf}
\alias{trunc_normal_cdf}
\title{Cumulative density function of a truncated normal distribution}
\usage{
trunc_normal_cdf(liab, a = stats::qnorm(0.05, lower.tail = F), b = Inf)
}
\arguments{
\item{liab}{true underlying liability for an individual}

\item{a}{the lower cutoff point of the truncated normal distribution}

\item{b}{the upper cutoff point of the truncated normal distribution}
}
\value{
Returns the estimated genetic liabilities.
}
\description{
Cumulative density function of a truncated normal distribution
}
\examples{
# See R/Example/example_nosib.R for an example of use and input.
# curve(trunc_normal_cdf, from = qnorm(0.05, lower.tail = F), to = 3.5, xname = "liab")
}
