% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{validate_proportion}
\alias{validate_proportion}
\title{Checking that proportions are valid}
\usage{
validate_proportion(prop, from_covmat = FALSE)
}
\arguments{
\item{prop}{A number, integer or numeric vector representing the proportions that
need to be validated.}

\item{from_covmat}{logical variable. Only used internally. allows for skip of negative check.}
}
\value{
If \code{prop} is a vector holding valid proportions of class \code{numeric}
or \code{integer} that are non-negative and at most one,
then the function will return TRUE. Otherwise, the function aborts.
}
\description{
\code{validate_proportion} checks whether proportions are valid, i.e.
whether they are non-negative and at most one.
}
\details{
This function can be used to check whether proportions are non-negative
and at most one.
}
\examples{
\dontrun{
validate_proportion(0.2)
validate_proportion(0.04)
validate_proportion(0)
validate_proportion(1)

# This will result in errors:
validate_proportion(2)
validate_proportion(-0.5)
}

}
