% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{validate_relatives}
\alias{validate_relatives}
\title{Checking that relatives are represented by valid strings}
\usage{
validate_relatives(relatives)
}
\arguments{
\item{relatives}{A string or character vector representing
the relatives.
All strings must be chosen among the following abbreviations
\itemize{
\item \code{g} (Genetic component of full liability)
\item \code{o} (Full liability)
\item \code{m} (Mother)
\item \code{f} (Father)
\item \code{c[0-9]*.[0-9]\*} (Children)
\item \code{mgm} (Maternal grandmother)
\item \code{mgf} (Maternal grandfather)
\item \code{pgm} (Paternal grandmother)
\item \code{pgf} (Paternal grandfather)
\item \code{s[0-9]*} (Full siblings)
\item \code{mhs[0-9]*} (Half-siblings - maternal side)
\item \code{phs[0-9]*} (Half-siblings - paternal side)
\item \code{mau[0-9]*} (Aunts/Uncles - maternal side)
\item \code{pau[0-9]*} (Aunts/Uncles - paternal side)
for the function to return TRUE.
}}
}
\value{
If \code{relatives} is a string or character vector such that
all strings are chosen from the mentioned list of strings,
then the function will return TRUE. Otherwise, the function is aborted.
}
\description{
\code{validate_relatives} checks whether relatives are represented
by valid abbreviations.
}
\details{
This function can be used to check whether relatives are represented
by valid abbreviations. A valid abbreviation is one of the following:
\itemize{
\item \code{g} (Genetic component of full liability)
\item \code{o} (Full liability)
\item \code{m} (Mother)
\item \code{f} (Father)
\item \code{c[0-9]*.[0-9]\*} (Children)
\item \code{mgm} (Maternal grandmother)
\item \code{mgf} (Maternal grandfather)
\item \code{pgm} (Paternal grandmother)
\item \code{pgf} (Paternal grandfather)
\item \code{s[0-9]*} (Full siblings)
\item \code{mhs[0-9]*} (Half-siblings - maternal side)
\item \code{phs[0-9]*} (Half-siblings - paternal side)
\item \code{mau[0-9]*} (Aunts/Uncles - maternal side)
\item \code{pau[0-9]*} (Aunts/Uncles - paternal side).
}
}
\examples{
\dontrun{
validate_relatives("g")
validate_relatives("o")
validate_relatives("mgm")

# This will result in errors:
validate_relatives("a")
validate_relatives(m)
}

}
