% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{check_valid_relatives}
\alias{check_valid_relatives}
\title{Checking that a relative is represented by a valid string}
\usage{
check_valid_relatives(relative)
}
\arguments{
\item{relative}{A string representing a relative.
The string must be one of the following
\itemize{
\item g (Genetic component of full liability)
\item o (Full liability)
\item m (Mother)
\item f (Father)
\item mgm (Maternal grandmother)
\item mgf (Maternal grandfather)
\item pgm (Paternal grandmother)
\item pgf (Paternal grandfather)
\item s[0-9]* (Full siblings)
\item mhs[0-9]* (Half-siblings - maternal side)
\item phs[0-9]* (Half-siblings - paternal side)
\item mau[0-9]* (Aunts/Uncles - maternal side)
\item pau[0-9]* (Aunts/Uncles - paternal side)
for the function to return TRUE.
}}
}
\value{
If \code{relative} is a string chosen from the mentioned list of strings,
then the function will return TRUE. Otherwise, the function is aborted.
}
\description{
\code{check_valid_relatives} checks whether a relative is represented
by a valid abbreviation.
}
\details{
This function can be used to check whether a relative is represented
by a valid abbreviation. A valid abbreviation is one of the following:
\itemize{
\item g (Genetic component of full liability)
\item o (Full liability)
\item m (Mother)
\item f (Father)
\item mgm (Maternal grandmother)
\item mgf (Maternal grandfather)
\item pgm (Paternal grandmother)
\item pgf (Paternal grandfather)
\item s[0-9]* (Full siblings)
\item mhs[0-9]* (Half-siblings - maternal side)
\item phs[0-9]* (Half-siblings - paternal side)
\item mau[0-9]* (Aunts/Uncles - maternal side)
\item pau[0-9]* (Aunts/Uncles - paternal side).
}
}
\examples{
check_valid_relatives("g")
check_valid_relatives("o")
check_valid_relatives("mgm")


\dontrun{
# This will result in errors:
get_relatedness("a")
get_relatedness(m)
}

}
