% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_matrix_functions.R
\name{construct_covmat}
\alias{construct_covmat}
\title{Constructing a covariance matrix for a variable number of
phenotypes}
\usage{
construct_covmat(
  fam_vec = c("m", "f", "s1", "mgm", "mgf", "pgm", "pgf"),
  n_fam = NULL,
  add_ind = TRUE,
  sq.herit = 0.5,
  phen_names = NULL
)
}
\arguments{
\item{fam_vec}{A vector of strings holding the different
family members. All family members must be represented by strings from the
following list:
\itemize{
\item m (Mother)
\item f (Father)
\item mgm (Maternal grandmother)
\item mgf (Maternal grandfather)
\item pgm (Paternal grandmother)
\item pgf (Paternal grandfather)
\item s[0-9]* (Full siblings)
\item mhs[0-9]* (Half-siblings - maternal side)
\item phs[0-9]* (Half-siblings - paternal side)
\item mau[0-9]* (Aunts/Uncles - maternal side)
\item pau[0-9]* (Aunts/Uncles - paternal side).
Defaults to c("m","f","s1","mgm","mgf","pgm","pgf").
}}

\item{n_fam}{A named vector holding the desired number of family members.
All names must be picked from the list mentioned above. Defaults to NULL.}

\item{add_ind}{A logical scalar indicating whether the genetic
component of the full liability as well as the full
liability for the underlying individual should be included in
the covariance matrix. Defaults to TRUE.}

\item{sq.herit}{Either a number representing the squared heritability
on liability scale for one single phenotype or a matrix holding the squared
heritability on liability scale for the different phenotypes as well as
the correlation between these phenotypes. When sq.herit is a matrix,
the heritabilities must be given on the diagonal,
while the off-diagonal entries must hold the correlation between phenotypes.
All squared heritabilities must be non-negative and at most 1.
All correlations must be between -1 and 1.
Defaults to 0.5.}

\item{phen_names}{A character vector holding the phenotype names. These names
will be used to create the row and column names for the covariance matrix.
If it is not specified, the names will default to pheno1, pheno2, etc.
Defaults to NULL.}
}
\value{
If either fam_vec or n_fam is used as the argument, if it is of
the required format and sq.herit is either a number satisfying
0 <= sq.herit <= 1 or a numeric matrix satisfying that all
diagonal entries are between 0 and 1 and that all off-diagonal
entries are between -1 and 1, then the output will be a named covariance matrix.
The number of rows and columns corresponds to the number of phenotypes times
the length of fam_vec or n_fam (+ 2 if add_ind=T).
If both fam_vec and n_fam are equal to c() or NULL, the function returns
a \eqn{(2 \times number of phenotypes) \times (2\times number of phenotypes)}
matrix holding only the correlation between the genetic component of the full
liability and the full liability for the underlying individual for all
phenotypes. If both fam_vec and n_fam are specified, the user is asked to
decide on which of the two vectors to use.
Note that the returned object has different attributes, such as
fam_vec, n_fam, add_ind, sq.herit and phenotype_names.
}
\description{
\code{construct_covmat} returns the covariance matrix for an
underlying individual and a variable number of its family members
for a variable number of phenotypes. It is a wrapper around
\code{\link{construct_covmat_single}} and \code{\link{construct_covmat_multi}}.
}
\details{
This function can be used to construct a covariance matrix for
a given number of family members. If sq.herit is a number,
each entry in this covariance matrix equals the percentage
of shared DNA between the corresponding individuals times
the squared heritability \eqn{h^2}. However, if sq.herit is a matrix,
each entry equals either the percentage of shared DNA between the corresponding
individuals times the squared heritability \eqn{h^2} or the
percentage of shared DNA between the corresponding individuals times
the correlation between the corresponding phenotypes. The family members
can be specified using one of two possible formats.
}
\examples{
construct_covmat()
construct_covmat(fam_vec = c("m","mgm","mgf","mhs1","mhs2","mau1"), n_fam = NULL, add_ind = TRUE, sq.herit = 0.5)
construct_covmat(fam_vec = NULL, n_fam = setNames(c(1,1,1,2,2), c("m","mgm","mgf","s","mhs")), add_ind = FALSE, sq.herit = 0.3)
construct_covmat(sq.herit = matrix(c(0.5,0.2,0.2,0.5), nrow = 2))

}
\seealso{
\code{\link{get_relatedness}}, \code{\link{construct_covmat_single}},
\code{\link{construct_covmat_multi}}
}
