% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_gen_liab.R
\name{estimate_gen_liability}
\alias{estimate_gen_liability}
\title{Estimate genetic liability based off of family history and prevalence information.}
\usage{
estimate_gen_liability(h2, phen, thr, ind = c(1), id_col = "ids", tol = 0.01)
}
\arguments{
\item{h2}{Liability scale heritability of the trait being analysed.}

\item{phen}{tibble or data.frame with the IDs and status of the genotyped individual and the parents and siblings.}

\item{thr}{tibble or data.frame with a row of each individual in the provided phen. Each row should contain the ID and threshold value needed for the model.}

\item{ind}{Indices to return from the gibbs sampler. 1 corresponds to the genetic liability of the first phenotype. c(1,5) corresponds to the genetic liability of the first two phenotypes provided, but with no siblings in the model.}

\item{id_col}{Column name for IDs of family members.}

\item{tol}{Convergence criteria of the gibbs sampler. Default is 0.01, meaning a standard error of the mean below 0.01}
}
\value{
Returns the estimated genetic liabilities.
}
\description{
Estimate genetic liability based off of family history and prevalence information.
}
\examples{
#See vignettes for example of use and input.


}
