% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_gen_liab_noFH.R
\name{estimate_gen_liability_noFH}
\alias{estimate_gen_liability_noFH}
\title{Estimate genetic liability similar to LT-FH}
\usage{
estimate_gen_liability_noFH(phen, h2, tol = 0.01)
}
\arguments{
\item{phen}{tibble or data.frame with status of the individual with columns ids, lower, and upper (thresholds).}

\item{h2}{Liability scale heritability of the trait being analysed.}

\item{tol}{Convergence criteria of the gibbs sampler. Default is 0.01, meaning a standard error of the mean below 0.01}
}
\value{
Returns the estimated genetic liabilities based on prevalence information.
}
\description{
Estimate genetic liability similar to LT-FH
}
\examples{
# See R/Example/example_nosib.R for an example of use and input.
}
