% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_genetic_liability.R
\name{estimate_liability}
\alias{estimate_liability}
\title{Estimating the genetic or full liability}
\usage{
estimate_liability(
  family,
  threshs,
  sq.herit = 0.5,
  pid = "PID",
  fam_id = "fam_ID",
  out = c(1),
  tol = 0.01,
  parallel = FALSE,
  always_add = c("g", "o")
)
}
\arguments{
\item{family}{A matrix, list or data frame that can be converted into a tibble.
Must have at least two columns that hold the family identifier and the corresponding
personal identifiers, respectively. That is, for each family in fam_id there should
be a list holding all individuals belonging to that family in pid. Note that the
personal identifiers for all individuals must have a special format, as it must end
with _$?$, where $ ? \in \{g,o,m,f,mgm,mgf,pgm,pgf,s[0-9]\ast, mhs[0-9]\ast,phs[0-9]\ast,
mau[0-9]\ast, pau[0-9]\ast \}$. See \code{\link{construct_covmat}} for more information.}

\item{threshs}{A matrix, list or data frame that can be converted into a tibble.
Must have at least three columns, one holding the personal identifier for all individuals,
and the remaining two holding the lower and upper thresholds, respectively.}

\item{sq.herit}{A number representing the squared heritability on liability scale
for a single phenotype. Must be non-negative and at most 1.
Defaults to 0.5.}

\item{pid}{A string holding the name of the column in \code{family} and
\code{threshs} that hold the personal identifier(s). Defaults to "PID".}

\item{fam_id}{A string holding the name of the column in \code{family} that
holds the family identifier. Defaults to "fam_ID".}

\item{out}{A character or numeric vector indicating whether the genetic component
of the full liability, the full liability or both should be returned. If out = c(1) or
out = c("genetic"), the genetic liability is estimated and returned. If out = c(2) or
out = c("full"), the full liability is estimated and returned. If out = c(1,2) or
out = c("genetic", "full"), both components are estimated and returned.
Defaults to c(1).}

\item{tol}{A number that is used as the convergence criterion for the Gibbs sampler.
Equals the standard error of the mean. That is, a tolerance of 0.2 means that the
standard error of the mean is below 0.2. Defaults to 0.01.}

\item{parallel}{A logical scalar indicating whether computations should be performed parallel.
In order for this to be possible, the user must install the library "future.apply" and create a plan
(see \code{\link[future.apply]{future_apply}}). Defaults to FALSE.}

\item{always_add}{A character vector or NULL. If always_add = c("g","o"), both the genetic component
of the full liability as well as the full liability will be added to the list of family members.
If always_add equals "g" or "o", the genetic component of the full liability or the full liability
will be added, respectively. If always_add = NULL, no component will be added.
Defaults to c("g","o").}
}
\value{
If family and threshs are two matrices, lists or data frames that can be converted into
tibbles, if family has two columns named like the strings represented in pid and fam_id, if
threshs has a column named like the string given in pid as well as a colimn named "lower" and
a column named "upper" and if the squared heritability, out, tol and always_add are of the required form,
then the function returns a tibble with either four or six columns (depending on the length of out).
The first two columns correspond to the columns fam_id and pid from family.
If out is equal to c(1) or c("genetic"), the third and fourth column hold the estimated genetic
liability as well as the corresponding standard error, respectively.
If out equals c(2) or c("full"), the third and fourth column hold the estimated full liability
as well as the corresponding standard error, respectively.
If out is equal to c(1,2) or c("genetic","full"), the third and fourth column hold the estimated
genetic liability as well as the corresponding standard error, respectively, while the fifth and
sixth column hold the estimated full liability as well as the corresponding standard error, respectively.
}
\description{
\code{estimate_liability} estimates the genetic component of the full
liability and/or the full liability for a number of individuals based
on their family history.
}
\details{
This function can be used to estimate either the genetic component of the
full liability, the full liability or both.
}
\examples{

sims <- simulate_under_LTM(fam_vec = c("m","f","s1"), n_fam = NULL, add_ind = T, sq.herit = 0.5, n_sim=500, pop_prev = .05)
estimate_liability(family = sims$fam_ID, threshs = sims$thresholds, sq.herit = 0.5, pid = "PID", fam_id = "fam_ID", out = c(1), tol = 0.01, parallel = FALSE, always_add = c("g","o"))
# 
sims <- simulate_under_LTM(fam_vec = c(), n_fam = NULL, add_ind = T, sq.herit = 0.5, n_sim=200, pop_prev = .05)
estimate_liability(family = sims$fam_ID, threshs = sims$thresholds, sq.herit = 0.5, pid = "PID", fam_id = "fam_ID", out = c("genetic"), tol = 0.01, parallel = FALSE, always_add = c("g","o"))

}
\seealso{
\code{\link[future.apply]{future_apply}}
}
