% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_gen_liab_multiple_traits.R
\name{estimate_liability_multi}
\alias{estimate_liability_multi}
\title{Estimating the genetic or full liability for multiple phenotypes}
\usage{
estimate_liability_multi(
  family,
  threshs,
  sq.herits,
  corrmat,
  pid = "PID",
  fam_id = "fam_ID",
  out = c(1),
  tol = 0.01,
  parallel = FALSE,
  progress = FALSE
)
}
\arguments{
\item{family}{A matrix, list or data frame that can be converted into a tibble.
Must have at least two columns that hold the family identifier and the corresponding
personal identifiers, respectively. That is, for each family in fam_id there should
be a list holding all individuals belonging to that family in pid. Note that the
personal identifiers for all individuals must have a special format. It must be end
with _?, where ? is one of the following abbreviations
\itemize{
\item g (Genetic component of the full liability)
\item o (full liability)
\item m (Mother)
\item f (Father)
\item mgm (Maternal grandmother)
\item mgf (Maternal grandfather)
\item pgm (Paternal grandmother)
\item pgf (Paternal grandfather)
\item s[0-9]* (Full siblings)
\item mhs[0-9]* (Half-siblings - maternal side)
\item phs[0-9]* (Half-siblings - paternal side)
\item mau[0-9]* (Aunts/Uncles - maternal side)
\item pau[0-9]* (Aunts/Uncles - paternal side).
See also \code{\link{construct_covmat}}.
}}

\item{threshs}{A matrix, list or data frame that can be converted into a tibble.
Must have at least five columns; one holding the personal identifier for all individuals,
and the remaining four holding the lower and upper thresholds for the first and second
phenotype, respectively. It must be possible to tie each pair of lower and upper thresholds
to a specific phenotype uniquely. This is done easily by adding _{name_of_phenotype} to
the column names lower and upper, e.g. lower_p1 and upper_p1 for the lower and upper
thresholds corresponding to the first phenotype.}

\item{sq.herits}{A numeric vector representing the squared heritability on liability scale
for all phenotypes. All entries in sq.herits must be non-negative and at most 1.}

\item{corrmat}{A numeric matrix holding the genetic correlation between the desired
number of phenotypes as well as the full correlation.
The full correlations must be given on the diagonal,
while the off-diagonal entries must hold the correlation between phenotypes.
All correlations must be between -1 and 1.}

\item{pid}{A string holding the name of the column in \code{family} and
\code{threshs} that hold the personal identifier(s). Defaults to "PID".}

\item{fam_id}{A string holding the name of the column in \code{family} that
holds the family identifier. Defaults to "fam_ID".}

\item{out}{A character or numeric vector indicating whether the genetic component
of the full liability, the full liability or both should be returned. If out = c(1) or
out = c("genetic"), the genetic liability is estimated and returned. If out = c(2) or
out = c("full"), the full liability is estimated and returned. If out = c(1,2) or
out = c("genetic", "full"), both components are estimated and returned.
Defaults to c(1).}

\item{tol}{A number that is used as the convergence criterion for the Gibbs sampler.
Equals the standard error of the mean. That is, a tolerance of 0.2 means that the
standard error of the mean is below 0.2. Defaults to 0.01.}

\item{parallel}{A logical scalar indicating whether computations should be performed parallel.
In order for this to be possible, the user must install the library "future.apply" and create a plan
(see \code{\link[future.apply]{future_apply}}). Defaults to FALSE.}

\item{progress}{A logical scalar indicating whether the function should display a progress bar.
Defaults to FALSE.}
}
\value{
If family and threshs are two matrices, lists or data frames that can be converted into
tibbles, if family has two columns named like the strings represented in pid and fam_id, if
threshs has a column named like the string given in pid as well as a column named "lower" and
a column named "upper" and if the squared heritability, out, tol and always_add are of the required form,
then the function returns a tibble with either four or six columns (depending on the length of out).
The first two columns correspond to the columns fam_id and pid from family.
If out is equal to c(1) or c("genetic"), the third and fourth column hold the estimated genetic
liability as well as the corresponding standard error, respectively.
If out equals c(2) or c("full"), the third and fourth column hold the estimated full liability
as well as the corresponding standard error, respectively.
If out is equal to c(1,2) or c("genetic","full"), the third and fourth column hold the estimated
genetic liability as well as the corresponding standard error, respectively, while the fifth and
sixth column hold the estimated full liability as well as the corresponding standard error, respectively.
}
\description{
\code{estimate_liability_multi} estimates the genetic component of the full
liability and/or the full liability for a number of individuals based
on their family history for a variable number of penotypes.
}
\details{
This function can be used to estimate either the genetic component of the
full liability, the full liability or both for a variable number of traits.
}
\seealso{
\code{\link[future.apply]{future_apply}}
}
