% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outdated_covariance_matrix_functions.R
\name{get_full_cov}
\alias{get_full_cov}
\title{Helps construct the correlation entries in the covariance matrix needed for LT-FH++ with multiple traits.}
\usage{
get_full_cov(corr_mat, n_sib = 0)
}
\arguments{
\item{corr_mat}{Matrix containing the heritabilities and correlations for each of the phenotypes provided. Heritabilities should be on the diagonal with the corresponing off-diagonal entry holding the correlation between two phenotypes.}

\item{n_sib}{Number of Siblings to include in the matrix.}
}
\value{
Returns the full covariance matrix based off of the heritabilities and corralations provided in corr_mat.
}
\description{
This function has been retired in favor of \code{\link{construct_covmat}}. This function is kept for legacy reasons.
}
\examples{
get_full_cov(matrix(.5, 2, 2))

}
