% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_matrix_functions.R
\name{get_relatedness}
\alias{get_relatedness}
\title{Relatedness between a pair of family members}
\usage{
get_relatedness(s1, s2, sq.herit = 0.5)
}
\arguments{
\item{s1, s2}{Strings representing the two family members.
The string must be one of the following:
\itemize{
\item g (Genetic component of full liability)
\item o (Full liability)
\item m (Mother)
\item f (Father)
\item mgm (Maternal grandmother)
\item mgf (Maternal grandfather)
\item pgm (Paternal grandmother)
\item pgf (Paternal grandfather)
\item s[0-9]* (Full siblings)
\item mhs[0-9]* (Half-siblings - maternal side)
\item phs[0-9]* (Half-siblings - paternal side)
\item mau[0-9]* (Aunts/Uncles - maternal side)
\item pau[0-9]* (Aunts/Uncles - paternal side).
}}

\item{sq.herit}{A number representing the squared heritability on liability scale.
Must be non-negative. Note that under the liability threshold model,
the squared heritability must also be at most 1.}
}
\value{
If both s1 and s2 are strings chosen from the mentioned list of strings and sq.herit is a number
satisfying 0 <= sq.herit, then the output will be a number that equals the percentage of shared
DNA between \code{s1} and \code{s2} times the squared heritability \code{sq.herit}.
}
\description{
\code{get_relatedness} returns the relatedness times the
squared heritability for a pair of family members
}
\details{
This function can be used to get the percentage of shared
DNA times the squared heritability \eqn{h^2} for two family members.
}
\note{
If you are only interested in the percentage of shared DNA, set \code{sq.herit = 1}.
}
\examples{
get_relatedness("g","o")
get_relatedness("g","f",sq.herit = 1)
get_relatedness("o","s",sq.herit = 0.3)


\dontrun{
# This will result in errors:
get_relatedness("a","b")
get_relatedness(m,mhs)
}

}
