% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFHPlus-package.R
\docType{package}
\name{LTFHPlus-package}
\alias{LTFHPlus}
\alias{LTFHPlus-package}
\title{LTFHPlus: Implementation of 'LT-FH++'}
\description{
Implementation of 'LT-FH++', an extension of 'LT-FH' using a Gibbs 
    sampler for sampling from the truncated multivariate normal distribution.
}
\author{
\strong{Maintainer}: Emil Michael Pedersen \email{emp@ph.au.dk}

Authors:
\itemize{
  \item Florian Privé \email{florian.prive.21@gmail.com} [thesis advisor]
}

Other contributors:
\itemize{
  \item Bjarni Jóhann Vilhjálmsson \email{bjv@econ.au.dk} [thesis advisor]
  \item Esben Agerbo \email{ea@econ.au.dk} [thesis advisor]
}

}
\keyword{internal}
