% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFHPlus-package.R
\docType{package}
\name{LTFHPlus-package}
\alias{LTFHPlus}
\alias{LTFHPlus-package}
\title{LTFHPlus: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\keyword{internal}
