% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFHPlus-package.R
\docType{package}
\name{LTFHPlus-package}
\alias{LTFHPlus}
\alias{LTFHPlus-package}
\title{LTFHPlus: Implementation of 'LT-FH++'}
\description{
Implementation of 'LT-FH++', an extension of 'LT-FH' which uses a Gibbs sampler for sampling from the truncated multivariate normal distribution. 'LT-FH++' allows for flexible family structures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EmilMiP/LTFHPlus}
}

}
\author{
\strong{Maintainer}: Emil Michael Pedersen \email{emp@ph.au.dk}

Authors:
\itemize{
  \item Florian Privé \email{florian.prive.21@gmail.com} [thesis advisor]
  \item Jette Steinbach \email{jst@econ.au.dk}
}

Other contributors:
\itemize{
  \item Bjarni Jóhann Vilhjálmsson \email{bjv@econ.au.dk} [thesis advisor]
  \item Esben Agerbo \email{ea@econ.au.dk} [thesis advisor]
  \item Lucas Rasmussen \email{lar.ncrr@au.dk} [contributor]
}

}
\keyword{internal}
