% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFHPlus-package.R
\docType{package}
\name{LTFHplus-package}
\alias{LTFHplus}
\alias{LTFHplus-package}
\title{LTFHplus: Extending the package 'LTFHplus'}
\description{
Extending the available package 'LTFHplus' in order to allow for more flexible family strctures. 'LT-FH++' is an extension of 'LT-FH' using a Gibbs sampler for sampling from the truncated multivariate normal distribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JetteS/LTFHPlus}
  \item Report bugs at \url{https://github.com/EmilMiP/LTFHPlus/issues}
}

}
\author{
\strong{Maintainer}: Jette Steinbach \email{jst@econ.au.dk}

Authors:
\itemize{
  \item Emil Michael Pedersen \email{emp@ph.au.dk}
  \item Florian Privé \email{florian.prive.21@gmail.com}
}

}
\keyword{internal}
